+++ Source/GUI/MemWatcher/Dialogs/DlgImportCTFile.cpp
@@ -111,7 +111,7 @@ bool DlgImportCTFile::willUseDolphinPointers() const
   return m_useDolphinPointers;
 }
 
-u64 DlgImportCTFile::getCommonBase() const
+u32 DlgImportCTFile::getCommonBase() const
 {
   return m_commonBase;
 }
